refactor: #115 - CommentList에서 comment 부분을 CommentItem 컴포넌트로 분리 추가
